clk optimization
Minimize the number of times the clk is set to external. Previously this would be done at the end of every clk_adjustment function call. Now with the added if statement, this is only done once for the first call to the clk_adjustment function (as required), for the clk must initially be set to internal.